Fixed unified errors
New builds of the Unified API broke compatibility. I’ve gone ahead and
fixed this throughout the app.

+++ StepCounter/Views/StepCounterController.cs
@@ -45,7 +45,7 @@ namespace StepCounter
             _stepManager.ForceUpdate();
         }
 
-        private void TodaysStepCountChanged(int stepCount)
+        private void TodaysStepCountChanged(nint stepCount)
         {
             //Setup Animation
             var stepCountAnimation = new CATransition ();
